+++ src/Exception.php
@@ -1,8 +1,0 @@
-<?php
-namespace JiNexus\Http;
-
-/**
- * Exception class
- */
-class Exception extends \Exception
-{ }