+++ src/HttpException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JiNexus\Http;
+
+use Exception;
+
+/**
+ * Exception class
+ */
+class HttpException extends Exception
+{ }